ledge refactor PLIP into separate component
